Deprecate jQuery datepicker and integrate react-datepicker
