+++ web/html/src/branding/css/susemanager/variables.less
@@ -93,6 +93,9 @@
 
 /* Input */
 @input-placeholder:  @gray;
+:root {
+  --input-selected: @eos-bc-green-500;
+}
 
 /* Modals */
 @modal-color:           @white;
