+++ web/html/src/branding/css/uyuni/variables.less
@@ -55,6 +55,11 @@
 @icon-legend:   lighten(@aside, 25%);
 @icon-legend-text: @text-color;
 
+/* Inputs */
+:root {
+  --input-selected: @spacewalk-green;
+}
+
 /* Roboto variable: font variable to be applied in different text elements*/
 @roboto-light: 'Roboto Light', sans-serif;
 @roboto-bold: 'Roboto Bold', sans-serif;
