+++ web/html/src/package.json
@@ -89,6 +89,7 @@
     "prop-types": "^15.7.2",
     "pwstrength-bootstrap": "1.2.6",
     "react": "^16.14.0",
+    "react-datepicker": "4.5.0",
     "react-dom": "^16.14.0",
     "react-hot-loader": "^4.13.0",
     "react-modal": "^3.13.1",
@@ -121,6 +122,7 @@
     "@types/lodash": "^4.14.168",
     "@types/node": "14.14.10",
     "@types/react": "^16.8.6",
+    "@types/react-datepicker": "^4.8.0",
     "@types/react-dom": "^16.8.6",
     "@types/react-modal": "^3.12.0",
     "@types/validator": "^13.1.3",
