+++ web/spacewalk-web.changes
@@ -1,3 +1,5 @@
+- Deprecate jQuery datepicker, integrate React datepicker
+
 -------------------------------------------------------------------
 Tue Feb 28 15:34:47 CET 2023 - jgonzalez@suse.com
